Move dynamic_decode from utils to decoders
